updated usbdux firmware from Bernd Porr:
Firmware now only measures as many channels as are in the channel list. If
there's only one channel in the list we can manage to measure within one
microframe. This makes it possible to measure with 8kHz max.
Added a counter. The counter is called every 2ms from an FX2 timer inerrupt. Low
priority so that measurements are not disturbed.
The bulk transfers are now two times faster (insn). Reading an A/D value or an
DIO takes now only 2ms or less. The trick is that the in endpoint is already
filled when the request is issued.
Fixed a bug in the FIFO resets.

Second counter did not count properly. That's fixed. Also
found two missing syncdelays.