patch from Ian Abbott <abbotti@mev.co.uk>:
I was running "comedi_test -v" on my new analog output driver and
noticed that it was getting stuck in a loop during the
"cmd_write_fast_1_chan" test.

It turned out that the test had written more data to the buffer than was
required to complete the acquisition (it was using cmd->stop_src =
TRIG_COUNT) and then got stuck in a loop waiting for the buffer to empty.

The attached patch stops the test_cmd_write_fast_1chan function writing
more data than is required to complete the acquisition.