configure: Allow choice of backend for PDF generation.
Add --with-pdf-backend=[yes|no|dblatex|fop|default] option which
defaults to "yes".  This controls the option passed to 'xmlto' to
generate PDFs.

If set to "yes" (the default) it looks for 'dblatex' and uses it as the
backend for 'xmlto' if found, otherwise 'xmlto' will use the default
backend.

If set to "no" (same as --without-pdf-backend), PDF generation is
disabled.

If set to "dblatex" it looks for 'dblatex' and uses it as the backend
for 'xmlto' if found, otherwise PDF generation is disabled.

If set to "fop" it looks for 'fop' and uses it as the backend
for 'xmlto' if found, otherwise PDF generation is disabled.

If set to "default", 'xmlto' will use the default backend.