doc: Use a stylesheet for dblatex
Add comedilib_dblatex_config.xsl stylesheet and pass it to dblatex when
producing PDF output.  The current version of the stylesheet doesn't
make much difference to the output, but it's somewhere we can put stuff
later.

We probably want to do something similar for the other PDF backends, but
those are currently broken anyway.