doc: Add Comedilib version number
In "comedilib.ent", Add an XML entity 'comedilib_version' expanding to
the Comedilib version '0.10.0'.  This is hard-edited in rather than set
dynamically.  Setting it dynamically would be nice but producing it from
an AC_OUTPUT() file isn't all that nice as it would end up in the
'build' tree and we'd some ugly 'make' rule to copy it to the 'source'
tree.  For now, just edit it manually as required.  At least having it
as an entity means it only needs to be set in a single documentation
file!

In "comedilib.xml", change the subtitle to mention Comedilib and the
version number, using the comediliv_version entity.