configure.ac: Expand rules for setting package/library version.
The comedilib manual version is usually only updated if the manual has
changed since the previous release (although it ought to change anyway
if any library interfaces have changed).  Allow a lower-case letter (or
sequence of letters) to be appended to the comedilib package version if
the only changes are outside the library sources and language bindings.