m4/am-check-python-headers.m4: Try python-config --includes
Some systems (e.g. Arch Linux) put the Python C headers in a strange
place, so try and run `$PYTHON-config --includes` first to find the -I
options to pass to the C compiler.  Fall back to the old method if it
fails.

Thanks to Darrell Enns for reporting the problem and suggesting a
solution.  I changed it a bit.