swig: remove comedi_ prefix from the function names
python always use the namespace of the library so it was redundant.

so what used to be written comedi.comedi_open() can be written comedi.open().

The main trick is to add a special python code that allows for the old names to still be compatible.
That avoids to break every single python program previously written.