lib/range.c: deal with samples > maxdata when converting to phys
By default, comedi_to_phys() and comedi_sample_to_phys() return NAN (not
a number) if the raw sample value is 0 or maxdata.  Change it to convert
any raw sample values above maxdata to NAN.  Such raw samples shouldn't
occur, but this seems like a sensible thing to do.  (Note that the
conversion of out-of-range values to NAN behavior can be changed via the
comedi_set_global_oor_behavior() function.)