Documents stop_arg!=0 when stop_src=TRIG_NONE
This patch documents the new interpretation of stop_arg when stop_src ==
TRIG_NONE.  In the new interpretation, if stop_src==TRIG_NONE, a user may
specify that a buffer of finite length (and less than or equal to the
maximum buffer size) be repeated indefinitely.

For now, only the NI MIO driver implements this new interpretation.  All
other drivers should continue to force stop_arg==0 when
stop_src==TRIG_NONE.  In this case, the old behavior is maintained.