Avoid using BIT(x) macro
The BIT(x) macro is used by some internal kernel headers and is not
used in user API kernel headers.  It was used incorrectly in the
"staging" kernel version of "comedi.h" in the "linux-next" tree for a
brief period.

Replace the use of BIT(x) with explicit masks.  Handle CR_INVERT
specially in SWIG to avoid it appearing as a negative integer.
_CR_FLAGS_MASK was defined twice, identically, in SWIG, so remove one of
the definitions.