swig/comedi.i: Fixed the "break backwards compatibility"
The lines:

%rename("%(strip:[COMEDI_])s", regextarget=1) "COMEDI_.*";
%rename("%(strip:[comedi_])s", regextarget=1) "comedi_.*";

caused errors on import due to _comedi function names not matching at
import time.  Stripping names is now completely done in the globals()
modification at the end, which works properly.

to indicate if the old names should be stripped or not.  This makes the
intent very clear in the generated comedi.py file.

[Reformatted commit message -- Ian Abbott]