comedi_config: Add --subdevice-buffer option
Add option --subdevice-buffer to allow buffer sizes for arbitrary
subdevices (that have asynchronous command support) to be configured.
This complements the --read-buffer and --write-buffer options.

The required argument is a comma-separated list alternating between
subdevice number and buffer size, and should end with a buffer size.
The option and its argument may be specified multiple times.