doc: Include drivers from Linux kernel sources in the comedilib manual
Split the "Kernel drivers" section into sub-sections for drivers built
from the Linux kernel sources, and the original comedi.org drivers.

Uses information from new file "doc/drivers-linux.txt" for information
about the drivers from the Linux kernel sources. This is built manually
using the scripts provided with the "comedi" driver sources.

The "doc/comedilib.ent" file now includes a 'linux_version' entity for
the version of the Linux kernel that the information in
"doc/drivers-linux.txt" was extracted from, and a 'comedi_date' entity
for the date the information in "doc/drivers.txt" was extracted from the
comedi drivers GIT repository.