lib/libinternal.h: Do some type checking in EXPORT_ALIAS_DEFAULT()
Make use of the GCC 'typeof' operator to make sure that the prototype of
the aliased function matches the prototype in "include/comedilib.h".

In "lib/buffer.c", need to declare comedi_get_front_count() because it
isn't declared in "comedilib.h" and is only being kept for backwards
compatibility.