comedi_test: Try to restore old buffer size or something sensible
The buffer size tests for each subdevice that supports async commands
currently leave the buffer set to maximum size if all goes well.

Try and restore the subdevice's buffer size to its original size before
the test if that was greater than 0 and less than the maximum size.  If
the original size was 0 and the maximum size is at least 10*4096, try
and set the buffer size to 10% of the maximum size, rounding up to a
multiple of 4096.

Also, don't try and set the buffer to maximum size if failed to get the
maximum size, or the maximum size is 0.