comedi_test: Change condition for setting buffer size to 10% of maximum
At the end of the buffer size tests for a subdevice that supports async
commands, if the original buffer size was 0, the buffer size is set to
10% of the maximum buffer size, rounded up to a multiple of 4096, if the
maximum buffer size is at least 10*4096.  Change the condition to
maxsize>4096, so that if the original buffer size was 0, the buffer size
will be set to the lesser of 4096 and 10% of the maximum size, rounded
up to a multiple of 4096.