swig/ruby: changes for building Ruby extension
Use AC_CONFIG_LINKS autoconf macro to add symbolic links to the build
directory for the setup.rb, ext/MANIFEST, and ext/extconf.rb files when
using separate source and build directories, and change the
swig/ruby/Makefile accordingly to create the comedi_ruby_wrap.c SWiG
wrapper locally in the build directory instead of polluting the source
directory.

When running setup.rb config, replace the --with-comedilib-dir= option
with separate --with-comedilib-include= and --with-comedilib-lib
options set so that it looks for the C header files in the source
directory, and looks for libcomedi.so in the build directory.  We
specify it to looks in the top-level, hidden lib/.libs directory where
the libcomedi.so file was actually built.  The old --with-comedilib-dir=
option didn't look in the hidden directory, so could only find a
libcomedi.so installed in the system directories.

Add a function call to swig/ruby/ext/extconf.rb to search for the
comedilib_version.h header file in the build directory, otherwise the
compilation of the comedi_ruby_wrap.c SWiG wrapper will fail when using
separate source and build directories.

In the comedi.i SWiG interface, for the SWIGRUBY parts, change the
'argout' typemap for the 'comedi_cmd' type to replace the call to
'output_helper()' with a call to 'SWIG_Ruby_AppendOutput()' to avoid a
compilation error due to the 'output_helper()' function not being
defined.  Also, do not defined the 'SWIG_USE_OLD_TYPEMAPS' macro.