configure.ac: Enable Ruby bindings by default
The building of the SWIG Ruby bindings was disabled by default since
April 2012 because they did not build cleanly from scratch, requiring an
existing installation of comedilib.  That is no longer the case.  Also,
the building of the Ruby bindings no longer pollutes the source
directory.  So re-enable them by default.  They can be disabled with the
--disable-ruby-bindings configure option.