demo/python: Change script shebang lines to run python3 instead of python
Modern distros tend to install the Python 3 executable with the name
'python3', and Python 2.7 has been end-of-life for a few years, so
change the shebang lines on the demo python scripts to run python3.

Also, in 'mmap.py', replace the call to the 'array.fromstring()' method
with 'array.frombytes()'.  The 'array.fromstring()' method was
deprecated in Python 3.2, and removed in Python 3.9.