doc: Build documentation in the build directory
When using separate source and build directories, build the
documentation in the build directory instead of the source directory.
This will ignore the built documentation files that were distributed in
the archive, which may be undesirable.

Allow the documentation to be installed even if not building the
documentation by changing the Makefile to install the documentation from
the source directory when documentation is not being built.