configure.ac: require --enable-docbook option to allow building documentation
Normal users probably will not want to (re-)build the documentation
shipped in the distribution archive, but because documentation is now
built in the build directory, that would be done if using separate
source and build directories.

Disable building the documentation by default.  The user can enable it
with the --enable-docbook option as long as the prerequisites for
building documenation are met.

Without the --enable-docbook option, the user can install the
documentation shipped in the source directory from the distribution
archive when doing `make install`.