README.Git: Add instructions for pre-building docs in the source directory
If it is desired to use separate source and build directories, but it is
not desired to build the documentation in the build directory, but it is
desired to install the documentation from the source directory, then it
is necessary to pre-build the documentation in the source directory
since the built documentation is not included in the git repository.

Add instructions to README.Git on how to build the documentation in the
source directory before configuring the build in the build directory.