doc/Makefile.am: Simplify dependencies and do more error checking
Rely on make's VPATH mechanism for determining whether dependencies are
located in the source or build directories.

Change rules that redirect a command output to the target filename to
redirect to a temporary filename (the target filename with a '~'
suffix), and only move the temporary file to the target file on success.