lib: Make comedi_dio_bitfield2() work better when INSN_BITS not supported
In comedi_dio_bitfield2() (and comedi_dio_bitfield()), if the subdevice
does not support the INSN_BITS instruction, then if the subdevice's
number of channels is less than base_channel + 32, then stop calling
comedi_dio_read() or comedi_dio_write() when base_channel+i would go
out of range, but always do at least 1 channel (the base_channel) in
order to catch errors.

This avoids the function always returning an error when the subdevice
does not support the INSN_BITS instruction and has less than
base_channel+32 channels, but it will still return an error if
base_channel is out of range.