configure.ac, Makefile.am: Add option to disable building and installing documentation
The new `--disable-doc` configure option will totally disable both the
building and installation of the API documentation (including the API
man-pages).  This is mainly for those who want to build and install
Comedilib from the Git repository sources, but do not want to install
the large dependencies for building the Comedilib manual and the API
man-pages.