doc/Makefile.am: Fix repeated command execution during parallel make
Multiple targets with the same rule resulted in commands being executed
repeatedly when make is running several jobs in parallel.

Use stamp-* files to serialize the rules that produce several target
files from the same prerequisites, and add these stamp-* files to the
distributed sources.  Also do this for the built XML files and remove
the BUILT_SOURCES directive to avoid having to build the XML files in a
sub-make.