doc/Makefile.am: Remove dupes from EXTRA_DIST and use MAINTAINERCLEANFILES
Both $(XML_IN) and $(BUILT_XML) were included in the expansion of
EXTRA_DIST, but $(BUILT_XML) is already included in the expansion of
XML_IN, so does not need to be included explicitly in the expansion of
EXTRA_DIST.

Define STATIC_XML_IN to list the static XML files, and make XML_IN
expand $(STATIC_XML_IN) and $(BUILT_XML).

Define BUILT_DIST to expand to all the built files that are to be
distributed.

Include $(BUILT_DIST) and $(STATIC_XML_IN) in the expansion of
EXTRA_DIST along with the other static files that were part of the
distribution.

Define MAINTAINERCLEANFILES to expand $(BUILT_DIST), and remove the
`maintainer-clean-local` rule.  `make maintainer-clean` will remove
any files listed in  MAINTAINERCLEANFILES.